+++ src/DynamoCore/Graph/Annotations/AnnotationModel.cs
@@ -7,6 +7,7 @@ using Dynamo.Graph.Notes;
 using Dynamo.Properties;
 using Dynamo.Utilities;
 using Newtonsoft.Json;
+using Autodesk.Workspaces;
 
 namespace Dynamo.Graph.Annotations
 {
