+++ src/DynamoCore/Graph/Nodes/NodeModel.cs
@@ -26,6 +26,7 @@ using String = System.String;
 using StringNode = ProtoCore.AST.AssociativeAST.StringNode;
 using System.Runtime.Serialization;
 using System.Diagnostics;
+using Autodesk.Workspaces;
 
 namespace Dynamo.Graph.Nodes
 {
